\name{EMMIXcskew-package}
\alias{EMMIXcskew-package}
\alias{EMMIXcskew}
\docType{package}
\title{
Finite mixture of multivariate canonical fundamental skew t-distributions  
}
\description{
Theis package implements an EM algorithm for 
fitting mixtures of multivariate canonical fundamental skew t (FM-CFUST) distributions.
Functions for random sample generation, initial value generation, 
and visualisation (in 2D and 3D) are also provided.
}
\details{
\tabular{ll}{
Package: \tab EMMIXcskew\cr
Type: \tab Package\cr
Version: \tab 0.9-4\cr
Date: \tab 2017-02-07\cr
Licence: \tab GPL\cr
LazyLoad: \tab yes\cr
}

}
\author{
S.X. Lee, G.J. McLachlan
}
\references{
Lee S.X. and McLachlan, G.J. (2016). Finite mixtures of canonical fundamental 
skew t-distributions: the unification of the restricted and unrestricted 
skew t-mixture models. Statistics and Computing 26, 573-589.

Lee S.X. and McLachlan, G.J. (2017). EMMIXcskew: An R Package for the 
Fitting of a Mixture of Canonical Fundamental Skew t-Distributions. 
Journal of Statistical Software 83(3), 1-32. 
URL \url{10.18637/jss.v083.i03}.   

}
\keyword{ package }
\seealso{
\code{\link{fmcfust}}, \code{\link{dfmcfust}}, \code{\link{rfmcfust}}, \code{\link{fmcfust.contour.3d}}
}
